pisource/requirements.txt
# pisource dependencies
# For Raspberry Pi Camera MJPEG Streaming

# Web server
flask>=2.3.0

# Picamera2 - install via apt on Raspberry Pi OS:
#   sudo apt install -y python3-picamera2
# Or via pip (may need additional system dependencies):
#   pip install picamera2
# Note: picamera2 is commented out because it should be installed
# via apt on Raspberry Pi OS for best compatibility
# picamera2>=0.3.12

# For dummy camera test mode (optional, only needed for --dummy)
numpy>=1.24.0
pillow>=9.5.0

